Move handle_messages to handlers_messages.cpp to satisfy File size gate
The #888-894 bugfix batch pushed handlers_chat.cpp to 820 code LOC, 20
over the 800 hard-review threshold. Relocate the handle_messages endpoint
(+ its new exception-envelope wrapper) into handlers_messages.cpp, next to
run_anthropic_stream_ which it drives — handlers_chat.cpp drops to 623 LOC,
handlers_messages.cpp to 708 (both under hard). No behavior change; the
moved code only references header-declared symbols.
