+++ tools/imp-server/handlers_messages.cpp
@@ -1,7 +1,8 @@
 // AUTO-SPLIT from handlers.cpp (verbatim move; see handlers_internal.h).
-// Anthropic /v1/messages streaming machinery: the AnthropicSSE event writer
-// and run_anthropic_stream_ (drives the real token loop, emits native
-// Anthropic SSE events). The handle_messages endpoint lives in handlers_chat.cpp.
+// Anthropic /v1/messages: the handle_messages endpoint (bottom of file) plus
+// its streaming machinery — the AnthropicSSE event writer and
+// run_anthropic_stream_ (drives the real token loop, emits native Anthropic
+// SSE events).
 
 #include "handlers.h"
 #include "handlers_internal.h"
@@ -638,4 +639,262 @@ bool run_anthropic_stream_(httplib::DataSink& sink, ChatRequestContext& ctx, Ser
     fprintf(stderr, "[%s] messages stream: %d prompt + %d completion tokens, %.1f ms (ttft=%.1f ms)\n",
             msg_id.c_str(), n_prompt_tokens, n_output_tokens, ms, ttft_ms);
     return true;
+}
+
+// ===========================================================================
+// Anthropic /v1/messages endpoint (moved here from handlers_chat.cpp to keep
+// that TU under the file-size gate; co-located with run_anthropic_stream_).
+// Non-streaming reuses the OpenAI path via a shim; streaming drives the real
+// per-token loop above.
+// ===========================================================================
+static void handle_messages_impl(const httplib::Request& req, httplib::Response& res, ServerState& state);
+
+void handle_messages(const httplib::Request& req, httplib::Response& res, ServerState& state) {
+    // Any exception escaping the impl — notably from the inner
+    // handle_chat_completions shim on the non-streaming path — must return the
+    // Anthropic error envelope ({"type":"error",...}), not the OpenAI-shaped one
+    // the global exception handler emits, which strict Anthropic SDK clients
+    // fail to parse (#891). res is untouched by the shim (it writes a separate
+    // shim_res) when the throw propagates, so it is safe to rewrite here.
+    try {
+        handle_messages_impl(req, res, state);
+    } catch (const std::exception& e) {
+        res.status = 500;
+        json err = {{"type", "error"}, {"error", {{"type", "server_error"}, {"message", e.what()}}}};
+        res.set_content(dump_safe(err), "application/json");
+    } catch (...) {
+        res.status = 500;
+        json err = {{"type", "error"}, {"error", {{"type", "server_error"}, {"message", "internal error"}}}};
+        res.set_content(dump_safe(err), "application/json");
+    }
+}
+
+static void handle_messages_impl(const httplib::Request& req, httplib::Response& res, ServerState& state) {
+    namespace anth = imp_server::anthropic;
+
+    // Capture original Anthropic request data for opt-in JSONL logging.
+    const auto t_log_start = std::chrono::system_clock::now();
+    const std::string log_endpoint = req.path;
+    std::string log_client_ip = req.get_header_value("X-Forwarded-For");
+    if (log_client_ip.empty())
+        log_client_ip = req.remote_addr;
+    const std::string log_raw_body = req.body;
+
+    json anth_body;
+    try {
+        anth_body = json::parse(req.body);
+    } catch (const std::exception& e) {
+        res.status = 400;
+        json err = {{"type", "error"},
+                    {"error",
+                     {{"type", "invalid_request_error"},
+                      {"message", std::string("Invalid JSON: ") + e.what()}}}};
+        res.set_content(dump_safe(err), "application/json");
+        return;
+    }
+
+    if (!anth_body.is_object()) {
+        res.status = 400;
+        json err = {{"type", "error"},
+                    {"error",
+                     {{"type", "invalid_request_error"}, {"message", "Request body must be a JSON object"}}}};
+        res.set_content(dump_safe(err), "application/json");
+        return;
+    }
+
+    // Anthropic requires max_tokens — if it's missing, supply a sane default
+    // matching the server's chat-completions default (handled downstream).
+    std::string anth_model = anth_body.value("model", "");
+    const bool want_stream = anth_body.value("stream", false);
+
+    // Transform -> OpenAI body.
+    json oai_body;
+    try {
+        oai_body = anth::anthropic_to_openai_body(anth_body);
+    } catch (const std::exception& e) {
+        res.status = 400;
+        json err = {{"type", "error"},
+                    {"error",
+                     {{"type", "invalid_request_error"},
+                      {"message", std::string("Failed to transform Anthropic body: ") + e.what()}}}};
+        res.set_content(dump_safe(err), "application/json");
+        return;
+    }
+
+    // ---- Real streaming path -------------------------------------------
+    // For stream=true we drive the same per-token batching-engine loop the
+    // OpenAI streaming path uses and emit native Anthropic SSE events as
+    // tokens arrive — TTFT is real first-token latency, not full-gen latency.
+    if (want_stream) {
+        // Build the chat request context from the transformed OpenAI body.
+        httplib::Request shim_req = req;
+        json shim_body = oai_body;
+        shim_body["stream"] = true;
+        shim_req.body = dump_safe(shim_body);
+        shim_req.headers.erase("Content-Length");
+        shim_req.headers.erase("content-length");
+
+        ChatRequestContext ctx;
+        g_in_anthropic_shim = true;  // suppress inner request-log (we log here)
+        bool ok = parse_chat_request_params(shim_req, res, state, ctx) &&
+                  snapshot_state_and_tokenize_(res, state, ctx);
+        g_in_anthropic_shim = false;
+        if (!ok) {
+            // parse/snapshot set an OpenAI-shaped error on res; re-wrap as
+            // an Anthropic error envelope.
+            json parsed;
+            try {
+                parsed = json::parse(res.body);
+            } catch (...) {
+                parsed = {{"error", {{"message", res.body}, {"type", "invalid_request_error"}}}};
+            }
+            json out = {{"type", "error"},
+                        {"error", parsed.value("error",
+                                               json{{"type", "invalid_request_error"}, {"message", "bad request"}})}};
+            res.set_content(dump_safe(out), "application/json");
+            return;
+        }
+
+        // Restore Anthropic logging context (parse_chat_request_params set
+        // these from the shim request; we log the outer Anthropic request).
+        ctx.log_skip = false;
+        ctx.log_endpoint = log_endpoint;
+        ctx.log_client_ip = log_client_ip;
+        ctx.log_raw_body = log_raw_body;
+        ctx.t_log_start = t_log_start;
+
+        // Vision is per-request now (req->image) and streams like any request.
+        auto imp_req = std::make_shared<imp::Request>();
+        imp_req->image = ctx.snap.vision_image;  // per-request vision (null for text)
+        imp_req->input_tokens = ctx.snap.tokens;
+        imp_req->max_tokens = ctx.params.max_tokens;
+        imp_req->temperature = ctx.params.temperature;
+        imp_req->top_p = ctx.params.top_p;
+        imp_req->top_k = ctx.params.top_k;
+        imp_req->seed = ctx.params.seed;
+        imp_req->pin_kv_prefix = ctx.params.cache_prompt;
+        imp_req->spec_ngram_override = ctx.params.spec_ngram_override;
+        // This is the streaming /v1/messages path — stay on per-step decode so
+        // SSE is real per-token rather than one burst at generation end (#754).
+        imp_req->stream = true;
+        imp_req->min_p = ctx.params.min_p;
+        imp_req->typical_p = ctx.params.typical_p;
+        imp_req->repetition_penalty = ctx.params.repetition_penalty;
+        imp_req->frequency_penalty = ctx.params.frequency_penalty;
+        imp_req->presence_penalty = ctx.params.presence_penalty;
+        imp_req->repeat_last_n = ctx.params.repeat_last_n;
+        imp_req->dry_multiplier = ctx.params.dry_multiplier;
+        imp_req->dry_base = ctx.params.dry_base;
+        imp_req->dry_allowed_length = ctx.params.dry_allowed_length;
+        imp_req->dry_penalty_last_n = ctx.params.dry_penalty_last_n;
+        imp_req->mirostat = ctx.params.mirostat;
+        imp_req->mirostat_tau = ctx.params.mirostat_tau;
+        imp_req->mirostat_eta = ctx.params.mirostat_eta;
+        imp_req->logprobs = ctx.params.req_logprobs;
+        imp_req->top_logprobs = ctx.params.top_logprobs;
+        imp_req->json_mode = ctx.params.json_mode;
+        imp_req->json_schema = ctx.params.json_schema_str;
+        imp_req->has_tools = ctx.params.has_tools;
+        imp_req->tpl_family = ctx.snap.tpl_family;
+        imp_req->logit_bias = ctx.params.logit_bias;
+        imp_req->think_budget = ctx.params.think_budget;
+        imp_req->status = imp::RequestStatus::PENDING;
+
+        auto server_req = std::make_shared<ServerRequest>();
+        server_req->request = imp_req;
+        {
+            std::lock_guard<std::timed_mutex> lock(state.mtx);
+            if (!state.batching || !state.batching->is_running()) {
+                res.status = 503;
+                json err = {{"type", "error"},
+                            {"error",
+                             {{"type", "server_error"}, {"message", "Inference engine not ready. Please retry."}}}};
+                res.set_content(dump_safe(err), "application/json");
+                return;
+            }
+            state.batching->submit(server_req);
+        }
+
+        std::string msg_id = anth::make_message_id(static_cast<uint64_t>(state.next_id.fetch_add(1)));
+        ctx.t_start = std::chrono::high_resolution_clock::now();
+
+        res.status = 200;
+        res.set_header("Cache-Control", "no-cache");
+        res.set_header("Connection", "keep-alive");
+        res.set_chunked_content_provider(
+            "text/event-stream",
+            [stream_ctx = std::move(ctx), &state, server_req, anth_model, msg_id](
+                size_t /*offset*/, httplib::DataSink& sink) mutable -> bool {
+                return run_anthropic_stream_(sink, stream_ctx, state, server_req, anth_model, msg_id);
+            });
+        return;
+    }
+
+    // ---- Non-streaming path: reuse the OpenAI handler via a shim --------
+    // httplib::Request is a plain struct, safe to copy. Force stream=false on
+    // the inner OpenAI call — we re-serialize the response as Anthropic JSON.
+    httplib::Request shim_req = req;
+    json shim_body = oai_body;
+    shim_body["stream"] = false;
+    shim_req.body = dump_safe(shim_body);
+    shim_req.headers.erase("Content-Length");
+    shim_req.headers.erase("content-length");
+
+    httplib::Response shim_res;
+    g_in_anthropic_shim = true;
+    handle_chat_completions(shim_req, shim_res, state);
+    g_in_anthropic_shim = false;
+
+    // Propagate error envelopes (transform them to Anthropic error shape).
+    // httplib::Response defaults status to -1 and auto-promotes to 200 only
+    // at send time; any other non-200 code set by handle_chat_completions is
+    // a real error we should forward.
+    const bool is_error = shim_res.status >= 400;
+    if (is_error) {
+        res.status = shim_res.status;
+        json parsed;
+        try {
+            parsed = json::parse(shim_res.body);
+        } catch (...) {
+            parsed = {{"error", {{"message", shim_res.body}, {"type", "server_error"}}}};
+        }
+        json out = {{"type", "error"},
+                    {"error", parsed.value("error", json{{"type", "server_error"}, {"message", "unknown"}})}};
+        res.set_content(dump_safe(out), "application/json");
+        return;
+    }
+
+    json oai_response;
+    try {
+        oai_response = json::parse(shim_res.body);
+    } catch (const std::exception& e) {
+        res.status = 500;
+        json err = {{"type", "error"},
+                    {"error",
+                     {{"type", "server_error"},
+                      {"message", std::string("Upstream returned non-JSON: ") + e.what()}}}};
+        res.set_content(dump_safe(err), "application/json");
+        return;
+    }
+
+    json anth_response = anth::openai_to_anthropic_response(oai_response, anth_model);
+
+    // JSONL log — built from Anthropic shapes so /v1/messages clients see
+    // exactly what they sent and what they got back.
+    {
+        auto t_end = std::chrono::system_clock::now();
+        double ms = std::chrono::duration<double, std::milli>(t_end - t_log_start).count();
+        int prompt_t = oai_response.value("usage", json::object()).value("prompt_tokens", 0);
+        int completion_t = oai_response.value("usage", json::object()).value("completion_tokens", 0);
+        std::string stop_reason = anth_response.value("stop_reason", "");
+        std::string req_id = anth_response.value("id", make_completion_id(state));
+        log_request_jsonl(state, /*skip=*/false, t_log_start, req_id, log_endpoint, log_client_ip,
+                          log_raw_body, ms, prompt_t, completion_t,
+                          stop_reason.empty() ? nullptr : stop_reason.c_str(), anth_response);
+    }
+
+    // Non-streaming requests are fully assembled above (the want_stream path
+    // returned earlier with a native incremental SSE stream).
+    res.status = 200;
+    res.set_content(dump_safe(anth_response), "application/json");
 }
